Create devcontainer
I created and configured a devcontainer that can be used to develop the
website and write new content. The devcontainer will provide a standard
environment for building and testing the website and new content. The
devcontainer will also allow me to work online using a GitHub Codespace
if I cannot work locally on my laptop.