Use SWA CLI to test the website
I configured the devcontainer to install Node.js 20.11.1 using NVM. I
created the .nvmrc file in the root directory of the repository to store
the Node.js version number for NVM to use when developing locally and
not in a devcontainer.

I created package.json, added the project metadata, and added a
development dependency on the Azure Static Web App CLI. Using the CLI
will allow me to somewhat test in a simulated environment. This isn't
so much important now, but in the future if I add any API Functions,
this will be useful for debugging and testing the APIs.