Fix responsive layout issues
I revised the CSS stylesheet to try to fix some layout issues I was
seeing on mobile devices. I cleaned up the CSS significantly and removed
some redundant and conflicting styles.

I changed the homepage to specify the cover image in the front matter
and add it as an embedded style attribute instead of making it static in
the CSS stylesheet. This will make it easier to update the home page
with a different image.