Add JetBrains Mono font for code blocks
I added the JetBrains Mono font to be downloaded from Google Fonts and
I switched from importing the fonts in the HTML HEAD element to
importing the fonts in the main.css file. I updated main.css to use the
JetBrains Mono font for code snippets and source code listings.