Add Hugo to the development container
I added Go and Hugo to the development container. I am using Hugo to
generate the static website that will be uploading to the hosting
environment. Hugo has a dependency on Go.

I also added the latest versions of Git and the GitHub CLI to the
development container. Installing Git gives me the latest Git features
and I will be using the GitHub CLI to automate certain actions or
interacting with the repository from the terminal.