Add Privacy Policy and Cookie Policy
I added support for the privacy and cookie policy and controller
solution from iubenda. I created partials for the privacy policy and
cookie policy links, and a partial template that renders the control
solution to display the banner when visiting the website.

I created the ci.yaml GitHub Actions workflow. This workflow will build
the website for pull requests and will identify any issues with the new
content.