Integrate Iubenda into the Website
I integrated the widgets and enforcement of the privacy and cookie
policies from Iubenda into the website. The widgets can be used to view
the privacy and cookie policy. The Iubenda Javascript code is included
to display the privacy and cookie policy information to new visitors and
to gain their consent.