Update Normal Theme Layout
I updated the layouts and CSS for the normal theme to render the website
more like how I'd like it to appear. I implemented the CSS to display a
cover photo as the background for the page header, and to render the
website title and main menu as a navigation bar at the top of the page.
I implemented a grid layout to center the content in the center of the
web page.