My Icebreaker Article
I wrote my first blog post. I wrote it following an icebreaker style to
introduce myself to new readers and tell them a bit about my past.

I integrated the widgets and enforcement of the privacy and cookie
policies from Iubenda into the website. The widgets can be used to view
the privacy and cookie policy. The Iubenda Javascript code is included
to display the privacy and cookie policy information to new visitors and
to gain their consent.

I updated the blog post template to include the Disqus comments section
so that readers can provide feedback.

I added the Google Analytics embedded template into the head.html
partial template. Using Google Analytics will allow me to track how
visitors use and visit the website and what content they are interested
in.

I added the OpenGraph metadata to the head of the normal template. This
will allow each page to publish metadata that will be used by search
engines and social media websites to display information about the
content in the page for hyperlink references and search results.

I added the Schema embedded template to the normal template. The Schema
template will output microdata about the page that can be used by search
engines, social media, or other applications to understand the content
of the page.

I added the X/Twitter Card embedded content template to the normal
theme. This will display a richer link to a page in the website when
referenced in an X/Twitter post.

I updated the layouts and CSS for the normal theme to render the website
more like how I'd like it to appear. I implemented the CSS to display a
cover photo as the background for the page header, and to render the
website title and main menu as a navigation bar at the top of the page.
I implemented a grid layout to center the content in the center of the
web page.