Create Comment Policy
I wrote a comment policy for my blog. The purpose of the comment policy
is to educate readers on what is and is not acceptable behavior when
making comments on my blog posts. The comment policy is intended to help
readers understand that I will be moderating comments and what I find
acceptable and not acceptable and will delete from my blog.