Write Article on Development Containers
I finished the article on building and using development containers.

I updated the list layout to show a breadcrumb and to allow the reader
to navigate to subsections at each section.

I updated the templates for blog posts. I added the summary text to the
top of the blog post and I added the last modified time to the bottom of
the blog post.

I updated the home page template and the list template to show the
publish time for blog post articles. I created the blog_post.html
partial template for both the home page template and the list template
to use.

I added the estimated reading time to the blog articles. I made some
text formatting changes.