Add License Terms to the Website
I added licensing terms for the content of the website, the website
source code, and source code published in articles. The actual content
of articles is licensed under a Creative Commons license. The source
code for the website and source code included in articles is being
released under the MIT License.