Run Website Locally in Development Container
I customized the command for `npm run dev` to add the `--host` flag,
which allows the development server to be accessible from outside the
container.

I updated the `devcontainer.json` file to include the necessary to
indicate that port 4321 should be forwarded from the container to
the host machine. This allows you to access the development server
running inside the container via `http://localhost:4321` on your host
machine.