Apply License to Project
I created LICENSE.md and added the text of the MIT license and the
CC BY-NC-ND 4.0 license. The MIT license applies to the source code for
the website and source code that is published on the website. The
CC BY-NC-ND 4.0 license applies to the content that is published on the
website.