Create Website Layout
I began creating the layout for the website. I created the CoverImage
component that will display a cover image at the top of the page. I
created the GlobalStyles component that can contain global styles for
the website and added styles to reset the default styles. I created
the BaseLayout component that will be used as the base layout for all
pages.