Publish the Website to GitHub Pages
I created a GitHub Actions workflow that will automatically build and 
publish the website to GitHub Pages when new content is added or the 
website structure is updated.

The workflow will be triggered by:

- pushes to the main branch
- manually executing the workflow
- every night at 1am MST

The scheduled trigger will help when there's content like blog posts 
that is published with a future release date.

The workflow will also run for pull requests, but it will only validate
that the website builds correctly.

Closes: #2