Create the Development Container
I created the configuration for the development container. Supporting
tools such as Visual Studio Code, GitHub Codespaces, and JetBrains IDEs
will use this file to build and run the development container. The
development container is the preferred environment for building and
testing the website. The development container is preconfigured will all
required tools and dependencies installed.